refactor(constant): enum 위치 변경
작업 내용
- 중요도에 따라 enum들이 따로 관리될 수 있도록
  enum들의 위치를 변경
