+++ src/main/java/lotto/controller/LottoController.java
@@ -2,9 +2,9 @@ package lotto.controller;
 
 import java.util.List;
 import lotto.config.ApplicationConfiguration;
-import lotto.constant.WinningTier;
 import lotto.domain.Lotto;
 import lotto.domain.WinningLotto;
+import lotto.domain.constant.WinningTier;
 import lotto.service.InputService;
 import lotto.service.LottoService;
 import lotto.service.OutputService;
