+++ src/main/java/lotto/view/constant/ExceptionMessage.java
@@ -1,4 +1,4 @@
-package lotto.constant;
+package lotto.view.constant;
 
 public enum ExceptionMessage {
     DUPLICATED_NUMBERS("중복되는 번호가 있습니다."),
