+++ src/main/java/lotto/view/constant/Prompt.java
@@ -1,4 +1,4 @@
-package lotto.constant;
+package lotto.view.constant;
 
 public enum Prompt {
     PURCHASE("구입금액을 입력해 주세요."),
